src/models/serve/requirements.txt
fastapi
uvicorn
pydantic
Pillow==10.0.0
pandas==2.0.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
numpy==1.24.3
nltk==3.8.1
bs4
python-multipart
python-dotenv
slowapi
python-jose[cryptography]
passlib[bcrypt]